feat: support for ILP/HTTP (#18)
This PR upgrades the client to support [ILP over HTTP transport](https://github.com/questdb/questdb/issues/4052) for data sent to QuestDB. This protocol provides a better dev-ex for data ingestion, including improved error messages, initialisation via configuration string, authentication, transactions, auto-flushing, and more.

+++ src/net-questdb-client/LineTcpSender.cs
@@ -6,7 +6,7 @@
  *    \__\_\\__,_|\___||___/\__|____/|____/
  *
  *  Copyright (c) 2014-2019 Appsicle
- *  Copyright (c) 2019-2022 QuestDB
+ *  Copyright (c) 2019-2024 QuestDB
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -33,27 +33,33 @@ using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Security;
 
+// ReSharper disable InconsistentNaming
+
 namespace QuestDB;
 
+/// <summary>
+///     Former ILP/TCP sender. Deprecated.
+/// </summary>
+[Obsolete("This has been superseded by ISender and the Sender factory.")]
 public class LineTcpSender : IDisposable
 {
     private static readonly RemoteCertificateValidationCallback AllowAllCertCallback = (_, _, _, _) => true;
     private static readonly long EpochTicks = new DateTime(1970, 1, 1).Ticks;
+    public static int DefaultQuestDbFsFileNameLimit = 127;
     private readonly BufferOverflowHandling _bufferOverflowHandling;
     private readonly List<(byte[] Buffer, int Length)> _buffers = new();
-    public static int DefaultQuestDbFsFileNameLimit = 127;
     private readonly Stream _networkStream;
-    private Socket? _underlyingSocket;
     private bool _authenticated;
     private int _currentBufferIndex;
     private bool _hasTable;
+    private int _lineStartBufferIndex;
+    private int _lineStartBufferPosition;
     private bool _noFields = true;
     private bool _noSymbols = true;
     private int _position;
     private bool _quoted;
     private byte[] _sendBuffer;
-    private int _lineStartBufferPosition;
-    private int _lineStartBufferIndex;
+    private Socket? _underlyingSocket;
 
     private LineTcpSender(Stream networkStream, int bufferSize,
         BufferOverflowHandling bufferOverflowHandling)
@@ -65,7 +71,10 @@ public class LineTcpSender : IDisposable
         QuestDbFsFileNameLimit = DefaultQuestDbFsFileNameLimit;
     }
 
-    /// <summary>Gets or sets a value, in milliseconds, that determines how long the underlying stream will attempt to write before timing out.</summary>
+    /// <summary>
+    ///     Gets or sets a value, in milliseconds, that determines how long the underlying stream will attempt to write
+    ///     before timing out.
+    /// </summary>
     public int WriteTimeout
     {
         get => _networkStream.WriteTimeout;
@@ -73,25 +82,25 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Shows if TCP socket is connected. Returns false if LineTcpSender is created from Stream
+    ///     Shows if TCP socket is connected. Returns false if LineTcpSender is created from Stream
     /// </summary>
     public bool IsConnected => _underlyingSocket?.Connected ?? false;
 
     /// <summary>
-    /// Closes the connection to QuestDB server and frees resources
+    ///     Maximum allowed column / table name. Usually set to 127 but can be overwritten in QuestDB server to higher value
+    /// </summary>
+    public int QuestDbFsFileNameLimit { get; set; }
+
+    /// <summary>
+    ///     Closes the connection to QuestDB server and frees resources
     /// </summary>
     public void Dispose()
     {
         _networkStream.Dispose();
     }
-    
-    /// <summary>
-    /// Maximum allowed column / table name. Usually set to 127 but can be overwritten in QuestDB server to higher value
-    /// </summary>
-    public int QuestDbFsFileNameLimit { get; set; }
 
     /// <summary>
-    /// Connects to QuestDB server
+    ///     Connects to QuestDB server
     /// </summary>
     /// <param name="host">QuestDB host name or IP address</param>
     /// <param name="port">QuestDB port name</param>
@@ -123,7 +132,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Uses connected Socket instance to create LineTcpSender
+    ///     Uses connected Socket instance to create LineTcpSender
     /// </summary>
     /// <param name="host">QuestDB server host name. Need if tslMode is Enabled</param>
     /// <param name="socket">TCP Socket instance. Socket must be connected to QuestDB ILP host</param>
@@ -161,6 +170,7 @@ public class LineTcpSender : IDisposable
                 };
                 await sslStream.AuthenticateAsClientAsync(options, cancellationToken);
                 if (!sslStream.IsEncrypted) throw new IOException("Cannot establish encrypted connection");
+
                 dataStream = sslStream;
             }
 
@@ -177,7 +187,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Creates LineTcpSender from IO Stream. Usually to be used with Network or SSL stream
+    ///     Creates LineTcpSender from IO Stream. Usually to be used with Network or SSL stream
     /// </summary>
     /// <param name="networkStream">Instance of Stream to operate on</param>
     /// <param name="bufferOverflowHandling">Initial size of IO buffer. Defaults to 4096</param>
@@ -191,7 +201,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Performs Key based Authentication with QuestDB
+    ///     Performs Key based Authentication with QuestDB
     /// </summary>
     /// <param name="keyId">Key or User Id</param>
     /// <param name="encodedPrivateKey">Base64 Url safe encoded Secp256r1 private key or `d` token in JWT key</param>
@@ -230,7 +240,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Set table name for the Line. Table name can be different from line to line.
+    ///     Set table name for the Line. Table name can be different from line to line.
     /// </summary>
     /// <param name="name">Table name</param>
     /// <returns>Itself</returns>
@@ -243,6 +253,7 @@ public class LineTcpSender : IDisposable
         if (!IsValidTableName(name))
         {
             if (IsEmpty(name)) throw new ArgumentException(nameof(name) + " cannot be empty");
+
             throw new ArgumentException(nameof(name) + " contains invalid characters");
         }
 
@@ -251,13 +262,13 @@ public class LineTcpSender : IDisposable
 
         _lineStartBufferIndex = _currentBufferIndex;
         _lineStartBufferPosition = _position;
-        
+
         EncodeUtf8(name);
         return this;
     }
 
     /// <summary>
-    /// Set value for a Symbol column. Symbols must be written before other columns
+    ///     Set value for a Symbol column. Symbols must be written before other columns
     /// </summary>
     /// <param name="symbolName">Symbol column name</param>
     /// <param name="value">Symbol value</param>
@@ -271,20 +282,22 @@ public class LineTcpSender : IDisposable
             if (!IsValidColumnName(symbolName))
             {
                 if (IsEmpty(symbolName)) throw new ArgumentException(nameof(symbolName) + " cannot be empty");
+
                 throw new ArgumentException(nameof(symbolName) + " contains invalid characters");
             }
-            
+
             Put(',').EncodeUtf8(symbolName).Put('=').EncodeUtf8(value);
             _noSymbols = false;
             return this;
         }
 
         if (!_hasTable) throw new InvalidOperationException("table expected");
+
         throw new InvalidOperationException("cannot write Symbols after Fields");
     }
 
     /// <summary>
-    /// Set value of String column.
+    ///     Set value of String column.
     /// </summary>
     /// <param name="name">Column name</param>
     /// <param name="value">Column value</param>
@@ -300,7 +313,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Set value of LONG column
+    ///     Set value of LONG column
     /// </summary>
     /// <param name="name">Column name</param>
     /// <param name="value">Column value</param>
@@ -312,7 +325,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Set value of BOOLEAN column
+    ///     Set value of BOOLEAN column
     /// </summary>
     /// <param name="name">Column name</param>
     /// <param name="value">Column value</param>
@@ -322,9 +335,9 @@ public class LineTcpSender : IDisposable
         Column(name).Put(value ? 't' : 'f');
         return this;
     }
-    
+
     /// <summary>
-    /// Set value of DOUBLE column
+    ///     Set value of DOUBLE column
     /// </summary>
     /// <param name="name">Column name</param>
     /// <param name="value">Column value</param>
@@ -336,7 +349,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Set value of TIMESTAMP column
+    ///     Set value of TIMESTAMP column
     /// </summary>
     /// <param name="name">Column name</param>
     /// <param name="timestamp">Column value</param>
@@ -347,22 +360,21 @@ public class LineTcpSender : IDisposable
         Column(name).Put(epoch / 10).Put('t');
         return this;
     }
-    
+
     /// <summary>
-    /// Finishes the line without specifying Designated Timestamp. QuestDB will set the timestamp at the time of writing to the table.
+    ///     Finishes the line without specifying Designated Timestamp. QuestDB will set the timestamp at the time of writing to
+    ///     the table.
     /// </summary>
     public void AtNow()
     {
         if (!_hasTable || (_noFields && _noSymbols))
-        {
             throw new InvalidOperationException("No symbols or column specified.");
-        }
 
         FinishLine();
     }
 
     /// <summary>
-    /// Finishes the line setting timestamp.
+    ///     Finishes the line setting timestamp.
     /// </summary>
     /// <param name="timestamp">Timestamp of the line</param>
     public void At(DateTime timestamp)
@@ -373,7 +385,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Finishes the line setting timestamp.
+    ///     Finishes the line setting timestamp.
     /// </summary>
     /// <param name="epochNano">Nanoseconds since Unix epoch</param>
     public void At(long epochNano)
@@ -381,9 +393,9 @@ public class LineTcpSender : IDisposable
         Put(' ').Put(epochNano);
         FinishLine();
     }
-    
+
     /// <summary>
-    /// Sends buffered lines to QuestDB
+    ///     Sends buffered lines to QuestDB
     /// </summary>
     public void Send()
     {
@@ -399,7 +411,7 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Sends buffered lines to QuestDB in async manner
+    ///     Sends buffered lines to QuestDB in async manner
     /// </summary>
     /// <param name="cancellationToken">Cancellation Token</param>
     public async Task SendAsync(CancellationToken cancellationToken = default)
@@ -416,37 +428,33 @@ public class LineTcpSender : IDisposable
     }
 
     /// <summary>
-    /// Cancel current unsent line. Works only in Extend buffer overflow mode.
+    ///     Cancel current unsent line. Works only in Extend buffer overflow mode.
     /// </summary>
     /// <exception cref="InvalidOperationException"></exception>
     public void CancelLine()
     {
         if (_bufferOverflowHandling == BufferOverflowHandling.SendImmediately)
-        {
             throw new InvalidOperationException("Cannot cancel line in BufferOverflowHandling.SendImmediately mode");
-        }
 
         _currentBufferIndex = _lineStartBufferIndex;
         _position = _lineStartBufferPosition;
     }
 
     /// <summary>
-    /// Frees unnecessary buffers. In BufferOverflowHandling.SendImmediately has no effect.
+    ///     Frees unnecessary buffers. In BufferOverflowHandling.SendImmediately has no effect.
     /// </summary>
     public void TrimExcessBuffers()
     {
-        int removeCount = _buffers.Count - _currentBufferIndex - 1;
-        if (removeCount > 0) 
-        {
-            _buffers.RemoveRange(_currentBufferIndex + 1, removeCount);
-        }
+        var removeCount = _buffers.Count - _currentBufferIndex - 1;
+        if (removeCount > 0) _buffers.RemoveRange(_currentBufferIndex + 1, removeCount);
     }
-    
+
     private static byte[] FromBase64String(string encodedPrivateKey)
     {
         var urlUnsafe = encodedPrivateKey.Replace('-', '+').Replace('_', '/');
         var padding = 3 - (urlUnsafe.Length + 3) % 4;
         if (padding != 0) urlUnsafe += new string('=', padding);
+
         return Convert.FromBase64String(urlUnsafe);
     }
 
@@ -486,7 +494,7 @@ public class LineTcpSender : IDisposable
         _noFields = true;
         _noSymbols = true;
     }
-    
+
     private LineTcpSender Column(ReadOnlySpan<char> columnName)
     {
         if (_hasTable)
@@ -494,9 +502,10 @@ public class LineTcpSender : IDisposable
             if (!IsValidColumnName(columnName))
             {
                 if (IsEmpty(columnName)) throw new ArgumentException(nameof(columnName) + " cannot be empty");
+
                 throw new ArgumentException(nameof(columnName) + " contains invalid characters");
             }
-            
+
             if (_noFields)
             {
                 Put(' ');
@@ -533,6 +542,7 @@ public class LineTcpSender : IDisposable
 
         var len = num.Length - pos;
         if (_position + len >= _sendBuffer.Length) NextBuffer();
+
         num.Slice(pos, len).CopyTo(_sendBuffer.AsSpan(_position));
         _position += len;
 
@@ -557,6 +567,7 @@ public class LineTcpSender : IDisposable
     {
         var l = tableName.Length;
         if (l > QuestDbFsFileNameLimit) return false;
+
         for (var i = 0; i < l; i++)
         {
             var c = tableName[i];
@@ -571,6 +582,7 @@ public class LineTcpSender : IDisposable
                         // Single dot allowed as compatibility,
                         // when someone uploads 'file_name.csv' the file name used as the table name
                         return false;
+
                     break;
                 case '?':
                 case ',':
@@ -614,6 +626,7 @@ public class LineTcpSender : IDisposable
     {
         var l = tableName.Length;
         if (l > QuestDbFsFileNameLimit) return false;
+
         for (var i = 0; i < l; i++)
         {
             var c = tableName[i];
@@ -676,6 +689,7 @@ public class LineTcpSender : IDisposable
             case ',':
             case '=':
                 if (!_quoted) Put('\\');
+
                 goto default;
             default:
                 Put(c);
